Crate rustc_const_math

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Rusty Mathematics

Note

This API is completely unstable and subject to change.

Reexports

pub use self::ConstInt::*;

pub use self::ConstUsize::*;

pub use self::ConstIsize::*;

Structs

ConstFloat 	 [
Experimental
]

Enums

ConstInt 	 [
Experimental
]

ConstIsize 	 [
Experimental
]
Depending on the target only one variant is ever used in a compilation. Anything else is an error. This invariant is checked at several locations

ConstMathErr 	 [
Experimental
]

ConstUsize 	 [
Experimental
]
Depending on the target only one variant is ever used in a compilation. Anything else is an error. This invariant is checked at several locations

Op 	 [
Experimental
]