Enum rustc_const_math::ConstInt

pub enum ConstInt {
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    Isize(ConstIsize),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    Usize(ConstUsize),
    Infer(u64),
    InferSigned(i64),
}

Unstable (rustc_private)

Variants

I8
Unstable (rustc_private)

I16
Unstable (rustc_private)

I32
Unstable (rustc_private)

I64
Unstable (rustc_private)

Isize
Unstable (rustc_private)

U8
Unstable (rustc_private)

U16
Unstable (rustc_private)

U32
Unstable (rustc_private)

U64
Unstable (rustc_private)

Usize
Unstable (rustc_private)

Infer
Unstable (rustc_private)

InferSigned
Unstable (rustc_private)

Methods

impl ConstInt

fn infer(self, other: Self) -> Result<(Self, Self), ConstMathErr>

Unstable (rustc_private)

If either value is Infer or InferSigned, try to turn the value into the type of the other value. If both values have no type, don't do anything

fn erase_type(self) -> Self

Unstable (rustc_private)

Turn this value into an Infer or an InferSigned

fn description(&self) -> &'static str

Unstable (rustc_private)

Description of the type, not the value

fn to_u64_unchecked(self) -> u64

Unstable (rustc_private)

Erases the type and returns a u64. This is not the same as -5i8 as u64 but as -5i8 as i64 as u64

fn to_u32(&self) -> Option<u32>

Unstable (rustc_private)

Converts the value to a u32 if it's in the range 0...std::u32::MAX

fn to_u64(&self) -> Option<u64>

Unstable (rustc_private)

Converts the value to a u64 if it's >= 0

fn is_negative(&self) -> bool

Unstable (rustc_private)

fn try_cmp(self, rhs: Self) -> Result<Ordering, ConstMathErr>

Unstable (rustc_private)

Compares the values if they are of the same type

fn wrap_incr(self) -> Self

Unstable (rustc_private)

Adds 1 to the value and wraps around if the maximum for the type is reached

fn int_type(self) -> Option<IntType>

Unstable (rustc_private)