Enum rustc_const_math::ConstInt

pub enum ConstInt {
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    I128(i128),
    Isize(ConstIsize),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    U128(u128),
    Usize(ConstUsize),
    Infer(u128),
    InferSigned(i128),
}

Unstable (rustc_private)

Variants
I8(i8)

Unstable (rustc_private)

I16(i16)

Unstable (rustc_private)

I32(i32)

Unstable (rustc_private)

I64(i64)

Unstable (rustc_private)

I128(i128)

Unstable (rustc_private)

Isize(ConstIsize)

Unstable (rustc_private)

U8(u8)

Unstable (rustc_private)

U16(u16)

Unstable (rustc_private)

U32(u32)

Unstable (rustc_private)

U64(u64)

Unstable (rustc_private)

U128(u128)

Unstable (rustc_private)

Usize(ConstUsize)

Unstable (rustc_private)

Infer(u128)

Unstable (rustc_private)

InferSigned(i128)

Unstable (rustc_private)

Methods

impl ConstInt

fn new_unsigned(val: u128, ty: UintTy, usize_ty: UintTy) -> Option<ConstInt>

Unstable (rustc_private)

Creates a new unsigned ConstInt with matching type while also checking that overflow does not happen.

fn new_signed(val: i128, ty: IntTy, isize_ty: IntTy) -> Option<ConstInt>

Unstable (rustc_private)

Creates a new unsigned ConstInt with matching type while also checking that overflow does not happen.

fn infer(self, other: Self) -> Result<(Self, Self), ConstMathErr>

Unstable (rustc_private)

If either value is Infer or InferSigned, try to turn the value into the type of the other value. If both values have no type, don't do anything

fn erase_type(self) -> Self

Unstable (rustc_private)

Turn this value into an Infer or an InferSigned

fn description(&self) -> &'static str

Unstable (rustc_private)

Description of the type, not the value

fn to_u128_unchecked(self) -> u128

Unstable (rustc_private)

Erases the type and returns a u128. This is not the same as -5i8 as u128 but as -5i8 as i128 as u128

fn to_u32(&self) -> Option<u32>

Unstable (rustc_private)

Converts the value to a u32 if it's in the range 0...std::u32::MAX

fn to_u64(&self) -> Option<u64>

Unstable (rustc_private)

Converts the value to a u64 if it's in the range 0...std::u64::MAX

fn to_u128(&self) -> Option<u128>

Unstable (rustc_private)

Converts the value to a u128 if it's in the range 0...std::u128::MAX

fn is_negative(&self) -> bool

Unstable (rustc_private)

fn try_cmp(self, rhs: Self) -> Result<Ordering, ConstMathErr>

Unstable (rustc_private)

Compares the values if they are of the same type

fn wrap_incr(self) -> Self

Unstable (rustc_private)

Adds 1 to the value and wraps around if the maximum for the type is reached

fn int_type(self) -> Option<IntType>

Unstable (rustc_private)